avance ajustes

+++ 1.Layers/1.4Domain/1.4.2SagasMainModule/Sagas.MainModule/ObjetosValor/Enum/enumIVA.cs
@@ -7,12 +7,12 @@ namespace Sagas.MainModule.ObjetosValor.Enum
 {
     public static class IvaEnum
     {
-        public static decimal p16 = (decimal)((int)enumIVA.iva16 / 100);
-        public static decimal p4 = (decimal)((int)enumIVA.iva4 / 100);
-        public static decimal p0 = (decimal)((int)enumIVA.iva0 / 100);
+        public static decimal p16 = (decimal)enumIVA.iva16 / 100;
+        public static decimal p4 = (decimal)enumIVA.iva4 / 100;
+        public static decimal p0 = (decimal)enumIVA.iva0 / 100;
     }
 
-    public enum enumIVA
+    public enum enumIVA 
     {
         iva16 = 16,
         iva4 = 4,
